+++ app/Http/Controllers/LeadsController.php
@@ -3,8 +3,18 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Http\UploadedFile;
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
+use Illuminate\Support\Facades\DB;
+
+use App\Http\Resources\Lead as ResourceLead;
 use App\Lead;
-use PDF;
+use App\File;
+use App\User;
+use MongoDB\BSON\ObjectID;
+use DataTables;
+use Gate;
 class LeadsController extends Controller
 {
     /**
@@ -12,13 +22,101 @@ class LeadsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $data['allLeads'] = Lead::all();
-   
+        abort_unless(Gate::any(['full_access','leads_show']), 404);
+
+        $data['title'] = 'Leads';
+
         return view('contents.leads', $data);
     }
 
+    public function all(Request $request){
+        abort_unless(Gate::any(['full_access','leads_show']), 404);
+
+        $leadModel = new Lead;
+        $columns = $request->columns;
+        $lastId = $request->last_id;
+        $totalData = Lead::count();
+        $totalFiltered = $totalData; 
+        $limit = intval($request->input('length'));
+        $start = intval($request->input('start'));
+        $order = $columns[intval($request->input('order.0.column'))];
+        $dir = $request->input('order.0.dir');
+
+        if(empty($request->input('search.value')))
+        {            
+            if($lastId == ''){
+                $leads = Lead::offset($start)
+                         ->limit($limit)
+                         ->orderBy($order['data'],$dir)
+                         ->get();
+            }else{
+                $leads = Lead::whereRaw([
+                    '_id' => ['$gt' => new ObjectID($lastId)]
+                ])->limit($limit)->orderBy($order['data'],$dir)->get();
+            }
+        }else {
+            $search = $request->input('search.value'); 
+            
+            $leads = Lead::orderBy($order['data'],$dir);
+
+            foreach($columns as $index=>$column){
+                if($leadModel->isInFillable($column['data'])){
+                    if($index == 0){
+                        $leads = $leads->where($column['data'], 'LIKE', $search.'%');
+                    }else{
+                        $leads = $leads->orWhere($column['data'], 'LIKE', $search.'%');
+                    }
+                    
+                }
+            }
+            $totalFiltered = $leads->get()->count();
+            $leads = $leads->limit($limit)->offset($start)->get();
+        }
+
+        $data = ResourceLead::collection($leads);
+        
+
+        $json_data = array(
+            "draw"            => intval($request->input('draw')),  
+            "recordsTotal"    => intval($totalData),  
+            "recordsFiltered" => intval($totalFiltered), 
+            "data"            => $data   
+        );
+
+        return response()->json($json_data);
+    }
+
+    public function uploadcsv(Request $request){
+        abort_unless(Gate::any(['full_access','leads_create']), 404);
+
+        if($request->hasFile('csv')){
+            if($request->file('csv')->isValid()){
+                // Get csv file
+                $file = $request->file('csv');
+                
+                // Make a image name based on user name and current timestamp
+                $filename = Str::slug(auth()->user()->name).'_'.time();
+                $extension = $file->getClientOriginalExtension();
+                $fileSize = $file->getSize();
+
+                $request->csv->storeAs('/public/files/csv', $filename.'.'.$extension);
+                $url = Storage::url('files/csv/'.$filename.'.'.$extension);
+            
+                $newfile = File::create([
+                    'user_id' => auth()->user()->id,
+                    'name' => $filename,
+                    'path' => $url,
+                    'extension' => $extension,
+                    'size' => $fileSize
+                ]);
+
+                if($newfile)
+                    return response()->json(array('success' => true, 'file_id' => $newfile->id));
+            }
+        }
+    }
     /**
      * Show the form for creating a new resource.
      *
@@ -80,9 +178,16 @@ class LeadsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Request $request)
+    public function destroy(Request $request, $id)
     {
-        $lead = Lead::find($request->id);
+        abort_unless(Gate::any(['full_access','leads_delete']), 404);
+
+        $user = User::find(auth()->user()->id);
+        if(!\Hash::check($request->password, $user->password)){
+            return response()->json(['success'=>false,'msg' => trans('passwords.invalid')]);
+        }
+
+        $lead = Lead::find($id);
         $lead->delete();
 
         if($lead){
@@ -92,10 +197,85 @@ class LeadsController extends Controller
         }
     }
 
-    public function import(){
+    public function importBody($id){
+        $file = File::find($id);
+        $body = array();
+        
+        if($file->status != 'imported'){
+            $filepath = storage_path('/app/public/files/csv/'.$file->name.'.'.$file->extension);
+            $file = fopen($filepath,"r");
 
+            $i = 0;
+                
+            while (($filedata = fgetcsv($file)) !== FALSE) {
+                $num = count($filedata);
+                
+                // Skip first row (Remove below comment if you want to skip the first row)
+                if($i == 0){
+                    for ($c=0; $c < $num; $c++) {
+                        $header[$c] = Str::slug($filedata[$c], '_');
+                    }
+                    $i++;
+                    continue; 
+                }
+
+                for ($c=0; $c < $num; $c++) {
+                    $body[$i][$header[$c]] = mb_convert_encoding($filedata[$c], 'UTF-8', 'UTF-8');
+                }
+
+                $i++;
+            }
+
+            fclose($file);
+        }
+
+        return response()->json($body);
+    }
+    public function import($id){
+        $file = File::find($id);
+        $leadModel = new Lead();
+        $headerOptions = $leadModel->getFillable();
+        $header = array();
+        
+        if($file->status != 'imported'){
+            $filepath = storage_path('/app/public/files/csv/'.$file->name.'.'.$file->extension);
+            $file = fopen($filepath,"r");
+
+            $i = 0;
+                
+            while (($filedata = fgetcsv($file)) !== FALSE) {
+                $num = count($filedata);
+                // Skip first row (Remove below comment if you want to skip the first row)
+                if($i == 0){
+                    for ($c=0; $c < $num; $c++) {
+                        $header[$c] = Str::slug($filedata[$c], '_');
+                    }
+                    $i++;
+                    break;
+                }
+                $i++;
+            }
+
+            fclose($file);
+        }
+
+        $data['title'] = 'Leads Import';
+        $data['header_options'] = $headerOptions;
+        $data['header'] = $header;
+        $data['file_id'] = $id;
+        //return response()->json($data);
+        return view('contents.leads-import', $data);
     }
 
+    public function importData(Request $request){
+        if(Lead::where('email',$request->email)->exists()){
+            return response()->json(array('success' => false, 'msg' => 'Data email address already exists in the database!'));
+        }else{
+            $lead = Lead::create($request->except('_token'));
+            return response()->json(array('success' => true, 'msg' => 'Data Imported Successfully'));
+        }
+        
+    }
     public function exportcsv(){
         $fileName = 'leads.csv';
         $leads = Lead::all();
